requirements.txt
playwright==1.58.0
pytest-playwright>=0.4.0
